+++ app/src/main/java/io/github/marcocipriani01/telescopetouch/renderer/ImageObjectManager.java
@@ -32,7 +32,7 @@ import io.github.marcocipriani01.telescopetouch.renderer.util.TextureManager;
 import io.github.marcocipriani01.telescopetouch.renderer.util.TextureReference;
 import io.github.marcocipriani01.telescopetouch.renderer.util.VertexBuffer;
 import io.github.marcocipriani01.telescopetouch.source.ImageSource;
-import io.github.marcocipriani01.telescopetouch.units.GeocentricCoordinates;
+import io.github.marcocipriani01.telescopetouch.astronomy.GeocentricCoordinates;
 
 /**
  * Manages the rendering of image objects.
@@ -91,9 +91,9 @@ public class ImageObjectManager extends RendererObjectManager {
             for (int i = 0; i < imageSources.size(); i++) {
                 ImageSource is = imageSources.get(i);
                 GeocentricCoordinates xyz = is.getLocation();
-                float px = xyz.x;
-                float py = xyz.y;
-                float pz = xyz.z;
+                float px = (float) xyz.x;
+                float py = (float) xyz.y;
+                float pz = (float) xyz.z;
 
                 float[] u = is.getHorizontalCorner();
                 float ux = u[0];
